Develop command clean code after merge

+++ Domain/BusinessObjects/Game.cs
@@ -13,7 +13,7 @@ namespace Domain.BusinessObjects
         public string ESRB {get; set;}
         public string Synopsis {get; set;}
         public string CoverPath {get; set;}
-        public int OverallScore {get; set;}
+        public float OverallScore {get; set;}
         public int Id { get; set; }
 
         public Game()
@@ -28,7 +28,6 @@ namespace Domain.BusinessObjects
         }
         public Game(string title, string genre, string esrb, string synopsis, string path,int id)
         {
-            //Owner = user;
             Title = title;
             Genre = genre;
             ESRB = esrb;
